requirements.txt
anyascii==0.3.1
asgiref==3.5.2
bcp47==0.0.4
beautifulsoup4==4.11.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.3
defusedxml==0.7.1
Django==4.1.3
django-allauth==0.51.0
django-debug-toolbar==3.7.0
django-enumchoicefield==3.0.1
django-extensions==3.2.1
django-filter==22.1
django-modelcluster==6.0
django-permissionedforms==0.1
django-recaptcha==3.0.0
django-taggit==3.1.0
django-treebeard==4.5.1
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
html5lib==1.1
idna==3.4
l18n==2021.3
oauthlib==3.2.2
openpyxl==3.0.10
Pillow==9.3.0
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
sentry-sdk==1.11.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
telepath==0.3
urllib3==1.26.12
wagtail==4.1.1
wagtail-django-recaptcha==1.0
wagtail-generic-chooser==0.4
wagtailsvg==0.0.35
wagtailvideos==4.0.1
webencodings==0.5.1
Willow==1.4.1
